refactor(app): split AlertEvaluationsTable into focused components
The evaluations table combined row rendering, expansion behavior, error
details, and infinite-scroll states in one 396-line file, over the repo's
300-line component guideline. Split by responsibility with no functionality
change:

- AlertStateBadge.tsx — the state badge (also folds the stateBadge()
  indirection into the component; it was aliased for export anyway)
- AlertEvaluationRow.tsx — the expandable per-window parent row, per-group
  child rows, and row-level formatting helpers
- AlertEvaluationsTable.tsx — the table shell, loading/empty/error states,
  and the infinite-scroll sentinel/retry affordance

AlertDetailPage now imports AlertStateBadge from its own module. Existing
unit tests exercise everything through the public AlertEvaluationsTable
component and pass unchanged.
